requirements.txt
telethon
pymongo
python-dotenv
PySide6
qasync
